Rename mpas_mesh_tools to just mpas_tools
It's too confusing to have mpas_mesh_tools next to mesh_tools.
Plus, the package is pulling in scripts from throughout the repo,
not just mesh tools.
