+++ docs/environment.yml
@@ -1,4 +1,4 @@
-name: mpas_mesh_tools_docs
+name: mpas_tools_docs
 channels:
   - conda-forge
 dependencies:
