+++ mpas_tools/__main__.py
@@ -5,22 +5,22 @@ MPAS mesh tools
 from __future__ import absolute_import, division, print_function, \
     unicode_literals
 
-import mpas_mesh_tools
+import mpas_tools
 
 import argparse
 
 
 def main():
     """
-    Entry point for the main script ``mpas_mesh_tools``
+    Entry point for the main script ``mpas_tools``
     """
 
     parser = argparse.ArgumentParser(
         description=__doc__, formatter_class=argparse.RawTextHelpFormatter)
     parser.add_argument('-v', '--version',
                         action='version',
-                        version='mpas_mesh_tools {}'.format(
-                                mpas_mesh_tools.__version__),
+                        version='mpas_tools {}'.format(
+                                mpas_tools.__version__),
                         help="Show version number and exit")
 
     args = parser.parse_args()
