--- visualization/paraview_vtk_field_extractor/mpas_mesh_tools
@@ -1,1 +1,0 @@
-../../mpas_mesh_tools/
